refactor: unify symbolic/concrete evaluators behind a Domain abstraction
bmc.py (symbolic, z3) and runtime.py (concrete, python) re-implemented the
same FSL expression semantics twice (~39 shared-name functions), a drift
hazard where the verifier and the replay Monitor could disagree.

Extract the semantically clean evaluator functions into a single shared core
in src/fslc/values.py, parameterized by a per-evaluator domain object
(bmc._SymDomain / runtime._ConcDomain) plus a recursive eval callback:
count, sum, quant, option compares, seq/struct compare, eval_is, eval_field,
eval_index, logical_map_access, plus 7 pure helpers. The domain encapsulates
the genuine differences (z3 vs python leaf ops, short-circuit vs eager
boolean fold, partial-op handling).

Deliberately NOT unified (genuinely divergent representations, not bugs):
_eval_seq_method/_eval_set_method/_apply_assign (z3.Store chains vs in-place
mutation), compute_updates (z3.If branch-merge vs concrete branch), and the
display helpers (z3-model extraction vs concrete read).

Behavior-preserving: the corpus verdict snapshot is byte-identical before and
after, and the evaluator-agreement test stays green.
